Start refactor
Refactoring so it matches existing components more closely.
