+++ src/App.vue
@@ -17,7 +17,7 @@ import HubStubCreateModal from "@/components/panels/edit/modals/HubStubCreateMod
 
 
 import ShelfListingModal from "@/components/panels/edit/modals/ShelfListing.vue";
-
+import AutoDeweyModal from "./components/panels/edit/modals/AutoDeweyModal.vue";
 import UpdateAvailableModal from "@/components/general/UpdateAvailableModal.vue";
 
 
@@ -41,6 +41,7 @@ export default {
     ShelfListingModal,
     DiacriticsConfigModal,
     UpdateAvailableModal,
+    AutoDeweyModal,
     TextMacroModal,
     NonLatinBulkModal,
     NonLatinAgentModal,
@@ -60,7 +61,7 @@ export default {
     ...mapStores(useConfigStore, useProfileStore, usePreferenceStore),
     // // gives read access to this.count and this.double
     ...mapState(useProfileStore, ['profilesLoaded', 'showValidateModal','profilesLoaded', 'showPostModal']),
-    ...mapWritableState(useProfileStore, ['showShelfListingModal','showHubStubCreateModal']),
+    ...mapWritableState(useProfileStore, ['showShelfListingModal','showHubStubCreateModal', 'showAutoDeweyModal']),
 
     ...mapState(usePreferenceStore, ['showPrefModal','catCode']),
     ...mapWritableState(usePreferenceStore, ['showLoginModal','showScriptshifterConfigModal','showDiacriticConfigModal','showTextMacroModal','showFieldColorsModal']),
@@ -86,10 +87,10 @@ export default {
     }
   },
 
-  async mounted() {    
-    console.log(this.configStore.versionMajor)    
+  async mounted() {
+    console.log(this.configStore.versionMajor)
 //     const configStore = useConfigStore()
-// const profileStore = useProfileStore()   
+// const profileStore = useProfileStore()
 
     this.preferenceStore.initalize()
     // this.profileStore.buildProfiles()
@@ -141,7 +142,7 @@ export default {
     <DiacriticsConfigModal v-model="showDiacriticConfigModal" />
   </template>
 
-  
+
   <template v-if="showShelfListingModal==true">
     <ShelfListingModal v-model="showShelfListingModal"  />
   </template>
@@ -151,7 +152,7 @@ export default {
   <template v-if="showTextMacroModal==true">
     <TextMacroModal v-model="showTextMacroModal"  />
   </template>
-  
+
   <template v-if="showNonLatinBulkModal==true">
     <NonLatinBulkModal v-model="showNonLatinBulkModal"  />
   </template>
@@ -159,16 +160,20 @@ export default {
   <template v-if="showNonLatinAgentModal==true">
     <NonLatinAgentModal v-model="showNonLatinAgentModal"  />
   </template>
-  
+
   <template v-if="showFieldColorsModal==true">
     <FieldColorsModal v-model="showFieldColorsModal"  />
   </template>
   <template v-if="showHubStubCreateModal==true">
     <HubStubCreateModal v-model="showHubStubCreateModal"  />
   </template>
-  
-  
-  
+
+  ??
+  <template v-if="showAutoDeweyModal==true">
+    <AutoDeweyModal v-model="showAutoDeweyModal"  />
+  </template>
+  !!
+
 </template>
