+++ src/components/panels/nav/Nav.vue
@@ -57,7 +57,7 @@
       ...mapState(usePreferenceStore, ['styleDefault', 'showPrefModal', 'panelDisplay']),
       ...mapState(useConfigStore, ['layouts']),
       ...mapWritableState(usePreferenceStore, ['showLoginModal','showScriptshifterConfigModal','showDiacriticConfigModal','showTextMacroModal','layoutActiveFilter','layoutActive','showFieldColorsModal']),
-      ...mapWritableState(useProfileStore, ['showPostModal', 'showShelfListingModal', 'activeShelfListData','showValidateModal', 'showRecoveryModal']),
+      ...mapWritableState(useProfileStore, ['showPostModal', 'showShelfListingModal', 'activeShelfListData','showValidateModal', 'showRecoveryModal', 'showAutoDeweyModal']),
       ...mapWritableState(useConfigStore, ['showNonLatinBulkModal','showNonLatinAgentModal']),
 
 
@@ -162,6 +162,11 @@
               this.activeShelfListData = {}
               this.showShelfListingModal = true
             }, icon:"🗄️" },
+
+            { text: "AutoDewey", click: () => {
+              this.deweyData = {}
+              this.showAutoDeweyModal = true
+            }, icon:"" },
 
             { is: 'separator'},
             {
