+++ src/stores/profile.js
@@ -99,6 +99,10 @@ export const useProfileStore = defineStore('profile', {
       componentGuid: null,
       componentPropertyPath:null
     },
+    showAutoDeweyModal: false,
+    deweyData: {
+      lcc: null,
+    },
 
     mostCommonNonLatinScript: null,
     nonLatinScriptAgents: {},
@@ -4540,7 +4544,76 @@ export const useProfileStore = defineStore('profile', {
 
     },
 
+    addDdc: async function(deweyInfo, guid, structure){
+      console.info("Add DDC: ", deweyInfo)
+      //Look to see if there is a DDC component
+      let activeProfile = this.profileStore.activeProfile
+      console.info(activeProfile)
+      let hasEmptyDDC = false
+      let ddcComponent = null
+      let lastClassifiction = null
+      for (let pt in activeProfile.rt["lc:RT:bf2:Monograph:Work"].pt){
+        if (pt.includes("id_loc_gov_ontologies_bibframe_classification__classification_numbers")){
+          const target = activeProfile.rt["lc:RT:bf2:Monograph:Work"].pt[pt]
+          console.info("target: ", target)
+          const userValue = target.userValue
+          console.info("userValue: ", userValue)
+          const classification = userValue["http://id.loc.gov/ontologies/bibframe/classification"][0]
+          const type = classification["@type"]
+          console.info("component: ", classification)
+          if (!pt.deleted){
+            lastClassifiction = pt
+            //type == "http://id.loc.gov/ontologies/bibframe/ClassificationDdc" &&
+            if (!Object.keys(classification).includes("http://id.loc.gov/ontologies/bibframe/classificationPortion")){
+              hasEmptyDDC = true
+              ddcComponent = classification
+            }
+          }
+        }
+      }
+      let newDDC
+      // if no empty ddc, create one
+      if (!hasEmptyDDC){
+        console.info("Creating component")
+        newDDC = await this.profileStore.duplicateComponentGetId(this.profileStore.returnStructureByComponentGuid(this.guid)['@guid'], this.structure, "lc:RT:bf2:Monograph:Work", lastClassifiction)
+        ddcComponent = activeProfile.rt["lc:RT:bf2:Monograph:Work"].pt[newDDC]
+      }
 
+      //add information to component
+      console.info("ddcComponent", ddcComponent)
+      let userValue = null
+      try{
+        userValue = ddcComponent.userValue["http://id.loc.gov/ontologies/bibframe/classification"][0]
+      } catch {
+        userValue = ddcComponent
+      }
+
+      let dewey = null
+      if (Object.keys(deweyInfo).includes("DDC")){
+        dewey = deweyInfo.DDC
+      } else {
+        dewey = deweyInfo.dewey
+      }
+
+      const newGuid = short.generate()
+      console.info("userValue: ", userValue)
+      userValue["@type"] = "http://id.loc.gov/ontologies/bibframe/ClassificationDdc"
+      userValue["http://id.loc.gov/ontologies/bibframe/classificationPortion"] = [{ "@guid": newGuid, "http://id.loc.gov/ontologies/bibframe/classificationPortion": String(dewey) }]
+
+      //Add the defaults:
+      // console.info("profile: ", this.profileStore.activeProfile.rt["lc:RT:bf2:Monograph:Work"].pt)
+      // const newComponent = activeProfile.rt["lc:RT:bf2:Monograph:Work"].pt[newDDC]
+      // const newStructure = this.profileStore.returnStructureByGUID(newComponent["@guid"])
+
+      // console.info("newComponent['@guid']: ", newComponent['@guid'])
+      // console.info("newStructure: ", newStructure)
+      // console.info("ddcComponent: ", ddcComponent)
+
+      // console.info("*/*/*/*/*/*/*/*/*/*/*/*/*/*/*/*/*/*/*/*/*/*/*/*/*/*/*/*/*/*/*/*")
+
+      // this.profileStore.insertDefaultValuesComponent(newComponent['@guid'], ddcComponent)
+
+    },
 
 
   },
